LayoutTests/platform/win-wk2/platform/win/plugins/window-geometry-initialized-before-set-window-expected.txt

CONSOLE MESSAGE: line 0: PLUGIN: ::MapWindowPoints failed

This tests that a plugin's HWND is already sized and positioned before NPP_SetWindow is called.
